pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Manohisoa <manohisoarajaonarivony@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
pyjwt = "^2.6.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"
daphne = "^4.0.0"
django-environ = "^0.9.0"
dj-database-url = "^1.0.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.990"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
